Add files via upload
This project takes a simple to-do list whether that be movies to watch, things to do around the house, or etc. and with a function gives the user a menu on the options for the to-do list to be changed according to what they want to do. Maybe they want to cross off a task, added one,  remove all the tasks completely, or just quit the program altogether this function allows them to do that and more.